More feedback

+++ kernel/include/bsp.h
@@ -33,10 +33,10 @@ pok_ret_t pok_bsp_irq_register(uint8_t irq, void (*handler)(void));
 
 void *pok_bsp_mem_alloc(size_t size);
 
-uint32_t pok_bsp_mem_base(void);
-uint32_t pok_bsp_mem_size(void);
-uint32_t pok_bsp_kernel_base(void);
-uint32_t pok_bsp_kernel_size(void);
+uintptr_t pok_bsp_mem_base(void);
+size_t pok_bsp_mem_size(void);
+uintptr_t pok_bsp_kernel_base(void);
+size_t pok_bsp_kernel_size(void);
 
 pok_ret_t pok_bsp_time_init();
